+++ src/lib.rs
@@ -117,59 +117,57 @@
 //! use akkorokamui::{api, client, Asset, Client, Response};
 //! use anyhow::{bail, Result};
 //! use serde::Deserialize;
-//! use std::{collections::HashMap, time::{Duration, SystemTime, UNIX_EPOCH}};
+//! use std::{
+//!     collections::HashMap,
+//!     time::{Duration, SystemTime, UNIX_EPOCH},
+//! };
 //!
 //! fn main() -> Result<()> {
-//!    let user_agent = "<product>/<product-version>";
-//!    let client: Client = client::with_user_agent(user_agent).build()?;
+//!     let user_agent = "<product>/<product-version>";
+//!     let client: Client = client::with_user_agent(user_agent).build()?;
 //!
-//!    #[derive(Debug, Deserialize)]
-//!    struct Trade {
-//!        price: String,
-//!        volume: String,
-//!        time: f32,
-//!        buy_sell: String,
-//!        market_limit: String,
-//!        miscellaneous: String,
-//!    }
+//!     #[derive(Debug, Deserialize)]
+//!     struct Trade {
+//!         price: String,
+//!         volume: String,
+//!         time: f32,
+//!         buy_sell: String,
+//!         market_limit: String,
+//!         miscellaneous: String,
+//!     }
 //!
-//!    #[derive(Debug, Deserialize)]
-//!    struct Trades {
-//!        #[serde(flatten)]
-//!        asset_pair_trades: HashMap<String, Vec<Trade>>,
-//!        last: String,
-//!    }
+//!     #[derive(Debug, Deserialize)]
+//!     struct Trades {
+//!         #[serde(flatten)]
+//!         asset_pair_trades: HashMap<String, Vec<Trade>>,
+//!         last: String,
+//!     }
 //!
-//!    let since = Duration::from_secs(10);
-//!    let since = match SystemTime::now().checked_sub(since) {
-//!        Some(since) => since.duration_since(UNIX_EPOCH)?.as_secs(),
-//!        _ => bail!("invalid duration"),
-//!    };
+//!     let since = Duration::from_secs(30 * 60);
+//!     let since = match SystemTime::now().checked_sub(since) {
+//!         Some(since) => since.duration_since(UNIX_EPOCH)?.as_secs(),
+//!         _ => bail!("invalid duration"),
+//!     };
 //!
-//!    // NOTE: the asset pair name may need to use the X and Z prefix depending
-//!    // on the Kraken classification system, where X stands for cryptocurrency
-//!    // based assets while Z is for fiat based assets. You can build a map of
-//!    // pair alternative name to asset pair effective name by querying all the
-//!    // AssetPairs from the homonymous API.
-//!    let asset_pair = Asset::XBT.pair(Asset::EUR);
-//!    let api = api::public::trades()
-//!        .with("pair", &asset_pair)
-//!        .with("since", since);
+//!     let asset_pair = Asset::XXBT.pair(Asset::ZEUR);
+//!     let api = api::public::trades()
+//!         .with("pair", &asset_pair)
+//!         .with("since", since);
 //!
-//!    let resp: Response<Trades> = client.send(api)?;
-//!    println!("{:?}", resp);
+//!     let resp: Response<Trades> = client.send(api)?;
+//!     println!("{:?}", resp);
 //!
-//!    if let Some(result) = resp.result {
-//!        if let Some(trades) = result.asset_pair_trades.get(&asset_pair) {
-//!            for trade in trades {
-//!                println!("price at {}: {}", trade.time, trade.price);
-//!            }
-//!        }
-//!   } else {
-//!      bail!("Cannot get trades: {:?}", resp.error);
-//!   }
+//!     if let Some(result) = resp.result {
+//!         if let Some(trades) = result.asset_pair_trades.get(&asset_pair) {
+//!             for trade in trades {
+//!                 println!("price at {}: {}", trade.time, trade.price);
+//!             }
+//!         }
+//!     } else {
+//!         bail!("Cannot get trades: {:?}", resp.error);
+//!     }
 //!
-//!    Ok(())
+//!     Ok(())
 //! }
 //! ```
 //!
@@ -188,9 +186,8 @@
 //! use anyhow::{bail, Result};
 //! use std::collections::HashMap;
 //!
-//! type Currency = String;
 //! type Amount = String;
-//! type Balance = HashMap<Currency, Amount>;
+//! type Balance = HashMap<Asset, Amount>;
 //!
 //! fn main() -> Result<()> {
 //!     let keys_path = "kraken.key";
@@ -206,7 +203,7 @@
 //!     println!("{:?}", resp);
 //!
 //!     if let Some(result) = resp.result {
-//!         println!("USD: {:?}", result.get(&Asset::USD.with_prefix()));
+//!         println!("GBP: {:?}", result.get(&Asset::ZGBP));
 //!     } else {
 //!         bail!("Cannot get balance: {:?}", resp.error);
 //!     }
@@ -227,58 +224,62 @@
 //! use std::collections::HashMap;
 //!
 //! fn main() -> Result<()> {
-//!    let keys_path = "kraken.key";
-//!    let credentials = Credentials::read(keys_path)?;
+//!     let keys_path = "kraken.key";
+//!     let credentials = Credentials::read(keys_path)?;
 //!
-//!    let user_agent = "<product>/<product-version>";
-//!    let client: Client = client::with_user_agent(user_agent)
-//!        .with_credentials(credentials)
-//!        .build()?;
+//!     let user_agent = "<product>/<product-version>";
+//!     let client: Client = client::with_user_agent(user_agent)
+//!         .with_credentials(credentials)
+//!         .build()?;
 //!
-//!    let asset_pairs = get_asset_pairs(&client)?;
-//!    let pair = Asset::XRP.pair(Asset::GBP);
-//!    let xrp_gbp = if let Some(name) = asset_pairs.get(&pair) {
-//!        name
-//!    } else {
-//!        bail!("{} asset pair name not found", pair)
-//!    };
+//!     let asset_pairs = get_asset_pairs(&client)?;
+//!     let pair = Asset::XXRP.pair(Asset::ZGBP);
+//!     let xrp_gbp = if let Some(name) = asset_pairs.get(&pair) {
+//!         name
+//!     } else {
+//!         bail!("{} asset pair name not found", pair)
+//!     };
 //!
-//!    let api = api::private::add_order()
-//!        // validate only, do not actually place any order
-//!        .with("validate", true)
-//!        .with("pair", &xrp_gbp)
-//!        .with("type", Order::Buy)
-//!        .with("ordertype", OrderType::TakeProfitLimit)
-//!        // take profit price trigger
-//!        .with("price", 0.19)
-//!        // limit price
-//!        .with("price2", 0.191)
-//!        .with("volume", 30)
-//!        // prefer fee in quote currency
-//!        .with("oflags", "fciq");
+//!     let api = api::private::add_order()
+//!         // validate only, do not actually place any order
+//!         .with("validate", true)
+//!         .with("pair", xrp_gbp)
+//!         .with("type", Order::Buy)
+//!         .with("ordertype", OrderType::TakeProfitLimit)
+//!         // take profit price trigger
+//!         .with("price", 0.19)
+//!         // limit price
+//!         .with("price2", 0.191)
+//!         .with("volume", 30)
+//!         // prefer fee in quote currency
+//!         .with("oflags", "fciq");
 //!
-//!    let resp: ResponseValue = client.send(api)?;
-//!    println!("{:?}", resp);
+//!     let resp: ResponseValue = client.send(api)?;
+//!     println!("{:?}", resp);
 //!
-//!    Ok(())
+//!     Ok(())
 //! }
 //!
 //! fn get_asset_pairs(client: &Client) -> Result<HashMap<String, String>> {
-//!    #[derive(Debug, Deserialize)]
-//!    struct AssetPair {
-//!        altname: String,
-//!    }
+//!     #[derive(Debug, Deserialize)]
+//!     struct AssetPair {
+//!         base: Asset,
+//!         quote: Asset,
+//!     }
 //!
-//!    type AssetPairs = HashMap<String, AssetPair>;
+//!     type AssetPairs = HashMap<String, AssetPair>;
 //!
-//!    let api = api::public::asset_pairs();
-//!    let resp: Response<AssetPairs> = client.send(api)?;
+//!     let api = api::public::asset_pairs();
+//!     let resp: Response<AssetPairs> = client.send(api)?;
 //!
-//!    if let Some(result) = resp.result {
-//!        Ok(result.into_iter().map(|(k, v)| (v.altname, k)).collect())
-//!    } else {
-//!       bail!("Cannot get asset pairs: {:?}", resp.error);
-//!    }
+//!     if let Some(result) = resp.result {
+//!         Ok(result
+//!             .into_iter()
+//!             .map(|(k, v)| (v.base.pair(v.quote), k))
+//!             .collect())
+//!     } else {
+//!         bail!("Cannot get asset pairs: {:?}", resp.error);
+//!     }
 //! }
 //! ```
 
@@ -318,7 +319,7 @@ mod tests {
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())
@@ -327,11 +328,11 @@ mod tests {
     #[test]
     fn assets_info() -> Result<()> {
         let client = Client::default();
-        let assets = [Asset::XBT, Asset::EUR, Asset::ETH];
+        let assets = [Asset::XXBT, Asset::ZEUR, Asset::XETH];
 
         let asset = assets
             .iter()
-            .map(|a| a.with_prefix())
+            .map(|a| a.to_string())
             .collect::<Vec<String>>()
             .join(",");
         let api = api::public::assets().with("asset", asset);
@@ -339,7 +340,7 @@ mod tests {
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())
@@ -349,13 +350,13 @@ mod tests {
     fn asset_pairs() -> Result<()> {
         let client = Client::default();
 
-        let asset_pair = Asset::XBT.pair(Asset::EUR);
+        let asset_pair = Asset::XXBT.pair(Asset::ZEUR);
         let api = api::public::asset_pairs().with("pair", &asset_pair);
         println!("{}", api);
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())
@@ -365,13 +366,13 @@ mod tests {
     fn ticker_info() -> Result<()> {
         let client = Client::default();
 
-        let asset_pair = Asset::XBT.pair(Asset::EUR);
+        let asset_pair = Asset::XXBT.pair(Asset::ZEUR);
         let api = api::public::ticker().with("pair", &asset_pair);
         println!("{}", api);
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())
@@ -381,13 +382,13 @@ mod tests {
     fn ohlc() -> Result<()> {
         let client = Client::default();
 
-        let asset_pair = Asset::XBT.pair(Asset::GBP);
+        let asset_pair = Asset::XXBT.pair(Asset::ZGBP);
         let api = api::public::ohlc().with("pair", &asset_pair);
         println!("{}", api);
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())
@@ -397,7 +398,7 @@ mod tests {
     fn depth() -> Result<()> {
         let client = Client::default();
 
-        let asset_pair = Asset::XBT.pair(Asset::GBP);
+        let asset_pair = Asset::XXBT.pair(Asset::ZGBP);
         let api = api::public::depth()
             .with("pair", &asset_pair)
             .with("count", 2);
@@ -405,7 +406,7 @@ mod tests {
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())
@@ -415,13 +416,13 @@ mod tests {
     fn trades() -> Result<()> {
         let client = Client::default();
 
-        let asset_pair = Asset::XBT.pair(Asset::USD);
+        let asset_pair = Asset::XXBT.pair(Asset::ZUSD);
         let api = api::public::trades().with("pair", &asset_pair);
         println!("{}", api);
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())
@@ -431,13 +432,13 @@ mod tests {
     fn spread() -> Result<()> {
         let client = Client::default();
 
-        let asset_pair = Asset::XBT.pair(Asset::USD);
+        let asset_pair = Asset::XXBT.pair(Asset::ZUSD);
         let api = api::public::spread().with("pair", &asset_pair);
         println!("{}", api);
 
         let resp: ResponseValue = client.send(api)?;
         println!("{:?}", resp);
-        assert!(resp.error.is_empty());
+        assert!(resp.is_success());
         assert!(resp.result.is_some());
 
         Ok(())